refactor(neon_news): Refactor package structure
