+++ packages/neon/neon_news/lib/src/blocs/articles.dart
@@ -1,4 +1,12 @@
-part of '../neon_news.dart';
+import 'dart:async';
+
+import 'package:neon/blocs.dart';
+import 'package:neon/models.dart';
+import 'package:neon/utils.dart';
+import 'package:neon_news/src/blocs/news.dart';
+import 'package:neon_news/src/options.dart';
+import 'package:nextcloud/news.dart' as news;
+import 'package:rxdart/rxdart.dart';
 
 enum FilterType {
   all,
