+++ packages/neon/neon_news/lib/src/dialogs/feed_update_error.dart
@@ -1,7 +1,10 @@
-part of '../neon_news.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:neon_news/l10n/localizations.dart';
+import 'package:nextcloud/news.dart' as news;
 
-class NewsFeedShowURLDialog extends StatefulWidget {
-  const NewsFeedShowURLDialog({
+class NewsFeedUpdateErrorDialog extends StatefulWidget {
+  const NewsFeedUpdateErrorDialog({
     required this.feed,
     super.key,
   });
@@ -9,31 +12,31 @@ class NewsFeedShowURLDialog extends StatefulWidget {
   final news.Feed feed;
 
   @override
-  State<NewsFeedShowURLDialog> createState() => _NewsFeedShowURLDialogState();
+  State<NewsFeedUpdateErrorDialog> createState() => _NewsFeedUpdateErrorDialogState();
 }
 
-class _NewsFeedShowURLDialogState extends State<NewsFeedShowURLDialog> {
+class _NewsFeedUpdateErrorDialogState extends State<NewsFeedUpdateErrorDialog> {
   @override
   Widget build(final BuildContext context) => AlertDialog(
-        title: Text(widget.feed.url),
+        title: Text(widget.feed.lastUpdateError!),
         actions: [
           ElevatedButton(
             onPressed: () async {
               await Clipboard.setData(
                 ClipboardData(
-                  text: widget.feed.url,
+                  text: widget.feed.lastUpdateError!,
                 ),
               );
               if (mounted) {
                 ScaffoldMessenger.of(context).showSnackBar(
                   SnackBar(
-                    content: Text(NewsLocalizations.of(context).feedCopiedURL),
+                    content: Text(NewsLocalizations.of(context).feedCopiedErrorMessage),
                   ),
                 );
                 Navigator.of(context).pop();
               }
             },
-            child: Text(NewsLocalizations.of(context).feedCopyURL),
+            child: Text(NewsLocalizations.of(context).feedCopyErrorMessage),
           ),
           ElevatedButton(
             onPressed: () {
