+++ packages/neon/neon_news/lib/src/dialogs/move_feed.dart
@@ -1,4 +1,8 @@
-part of '../neon_news.dart';
+import 'package:flutter/material.dart';
+import 'package:neon/widgets.dart';
+import 'package:neon_news/l10n/localizations.dart';
+import 'package:neon_news/src/widgets/folder_select.dart';
+import 'package:nextcloud/news.dart' as news;
 
 class NewsMoveFeedDialog extends StatefulWidget {
   const NewsMoveFeedDialog({
