+++ packages/neon/neon_news/lib/src/options.dart
@@ -1,4 +1,8 @@
-part of 'neon_news.dart';
+import 'package:neon/platform.dart';
+import 'package:neon/settings.dart';
+import 'package:neon/sort_box.dart';
+import 'package:neon_news/l10n/localizations.dart';
+import 'package:neon_news/src/blocs/articles.dart';
 
 class NewsAppSpecificOptions extends NextcloudAppOptions {
   NewsAppSpecificOptions(super.storage) {
