+++ packages/neon/neon_news/lib/src/pages/feed.dart
@@ -1,4 +1,8 @@
-part of '../neon_news.dart';
+import 'package:flutter/material.dart';
+import 'package:neon_news/src/blocs/articles.dart';
+import 'package:neon_news/src/blocs/news.dart';
+import 'package:neon_news/src/widgets/articles_view.dart';
+import 'package:nextcloud/news.dart' as news;
 
 class NewsFeedPage extends StatelessWidget {
   const NewsFeedPage({
