+++ packages/neon/neon_news/lib/src/sort/articles.dart
@@ -1,4 +1,6 @@
-part of '../neon_news.dart';
+import 'package:neon/sort_box.dart';
+import 'package:neon_news/src/options.dart';
+import 'package:nextcloud/news.dart' as news;
 
 final articlesSortBox = SortBox<ArticlesSortProperty, news.Article>(
   {
