+++ packages/neon/neon_news/lib/src/widgets/feeds_view.dart
@@ -1,4 +1,18 @@
-part of '../neon_news.dart';
+import 'package:flutter/material.dart';
+import 'package:neon/blocs.dart';
+import 'package:neon/sort_box.dart';
+import 'package:neon/utils.dart';
+import 'package:neon/widgets.dart';
+import 'package:neon_news/l10n/localizations.dart';
+import 'package:neon_news/src/blocs/news.dart';
+import 'package:neon_news/src/dialogs/feed_show_url.dart';
+import 'package:neon_news/src/dialogs/feed_update_error.dart';
+import 'package:neon_news/src/dialogs/move_feed.dart';
+import 'package:neon_news/src/options.dart';
+import 'package:neon_news/src/pages/feed.dart';
+import 'package:neon_news/src/sort/feeds.dart';
+import 'package:neon_news/src/widgets/feed_icon.dart';
+import 'package:nextcloud/news.dart' as news;
 
 class NewsFeedsView extends StatelessWidget {
   const NewsFeedsView({
